LPS-193341 Moves logic to LayoutSetPrototypeHelper
