+++ modules/apps/layout/layout-set-prototype-api/src/main/java/com/liferay/layout/set/prototype/helper/LayoutSetPrototypeHelper.java
@@ -31,4 +31,8 @@ public interface LayoutSetPrototypeHelper {
 			String friendlyURL)
 		throws PortalException;
 
+	public void resetPrototype(Layout layout) throws PortalException;
+
+	public void resetPrototype(LayoutSet layoutSet) throws PortalException;
+
 }
