+++ portal-kernel/src/com/liferay/sites/kernel/util/Sites.java
@@ -114,10 +114,6 @@ public interface Sites {
 	public void removeMergeFailFriendlyURLLayouts(LayoutSet layoutSet)
 		throws PortalException;
 
-	public void resetPrototype(Layout layout) throws PortalException;
-
-	public void resetPrototype(LayoutSet layoutSet) throws PortalException;
-
 	public void setMergeFailCount(
 			LayoutPrototype layoutPrototype, int newMergeFailCount)
 		throws PortalException;
